Add Git mail mapping
Added a Git mailmap (1) file to link to in documentations to allow
contributors to send mails regarding security issues.
This prevents unnecessary overhead of updating all documents when new
core team and members and contributors are added and additionally adds
the main functionality of the file: Mapping commits when someone uses a
different email address.

References:
  (1) https://git-scm.com/docs/git-shortlog#_mapping_authors

Epic: GH-8
Resolves GH-10