Loop on TypeScript configuration override
In GH-32 [1] the `@arcticicestudio/eslint-config-typescript` package
migrated to the latest plugin versions and added the `override`
field [1] to its index to simplify the usage for consumers by removing
the need to explicitly define it per project. Unfortunately this
resulted in a crash loop when doing so and also blocks users from
overriding single rules based on the project needs.

To fix this problem, the `override` field has been removed again,
leaving it up to the user to define and customize the ESLint
configuration for TypeScript to fit the project needs.

[1]: https://github.com/arcticicestudio/styleguide-javascript/pull/32
[2]: https://github.com/arcticicestudio/styleguide-javascript/blob/ab59f71f/packages/%40arcticicestudio/eslint-config-typescript/index.js#L31-L36

Co-authored-by: Sven Greb <development@svengreb.de>

GH-50