`svengreb` GitHub account and `@svengreb` npm package scope migration
With the retirement of the "Arctic Ice Studio" personal & Nord project
brand [1] this project has also moved to the "real-in-person" identity
"Sven Greb" both in the context of the repository to the `svengreb`
GitHub account [2] and the `@svengreb` npm package scope.
During the migration the current npm packages
`@arcticicestudio/eslint-config-base` [3],
`@arcticicestudio/eslint-config`[8] and
`@arcticicestudio/eslint-config-typescript` [9] were deprecated in favor
of the new and upcoming `@svengreb/eslint-config-base`[10],
`@svengreb/eslint-config` [11] and
`@svengreb/eslint-config-typescript` [12] packages that will be
published afterwards.

Also the current visual representation of this style guide [6] through
the way too outdated and deprecated GitBook [5] major version `2` has
been unpublished and removed. The documentations and references have
been updated to use the GitHub repository with the Markdown rendering
instead for now until a custom website has been implemented using a
modern "TechStack" like Next.js [7].

[1]: https://github.com/orgs/nordtheme/discussions/183#retire-arctic-ice-studio-as-nord-brand
[2]: https://github.com/svengreb
[3]: https://www.npmjs.com/package/@arcticicestudio/eslint-config-base
[4]: https://www.npmjs.com/package/@svengreb/remark-preset-lint
[5]: https://www.gitbook.com
[6]: https://arcticicestudio.github.io/styleguide-markdown
[7]: https://nextjs.org
[8]: https://www.npmjs.com/package/@arcticicestudio/eslint-config-base
[9]: https://www.npmjs.com/package/@arcticicestudio/eslint-config-typescript
[10]: https://www.npmjs.com/package/@svengreb/eslint-config-base
[11]: https://www.npmjs.com/package/@svengreb/eslint-config
[12]: https://www.npmjs.com/package/@svengreb/eslint-config-typescript

GH-88